requirements.txt
opencv-contrib-python
controlnet_aux
diffusers
torch
torchvision
requests
safetensors.torch
transformers
